fix: close colony menu on action button push
When on mobile or generally small-ish screens the colony menu stayed open when clicking the "Create action" button. This resulted in a state where seemingly nothing happened as the colony menu was covering the action sidebar. This fixes that by closing the colony menu before the action sidebar is opened. This also applies for all the "related actions" that now close the menu as well, when clicked.